requirements.txt
accelerate==0.19.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
altair==5.4.0
analytics-python==1.4.post1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
auto_gptq==0.7.1
backoff==1.10.0
baukit @ git+https://github.com/davidbau/baukit@9d51abd51ebf29769aecc38c4cbef459b731a36e
bcrypt==4.2.0
beautifulsoup4==4.12.3
benepar==0.2.0
bigmodelvis==0.0.1
bitermplus==0.7.0
bitsandbytes==0.39.0
blis==1.0.1
braceexpand==0.1.7
brokenaxes==0.6.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.18.1
cmake==3.29.6
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
confection==0.1.5
contourpy==1.2.1
crcmod==1.7
cryptography==43.0.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dataclasses-json==0.6.7
datasets==2.20.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1728594112391/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
decord==0.6.0
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
einops-exts==0.0.4
email_validator==2.2.0
en_core_web_sm @ file:///home/mlsnrs/data/wyn/CLIP/MMVP/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.4.0
filelock==3.15.4
flash-attention==1.0.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.3
gdown==5.2.0
gekko==1.2.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
gradio==3.40.0
gradio_client==0.7.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.12.0
httpcore==0.15.0
httplib2==0.22.0
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.0
inflection==0.5.1
iopath==0.1.10
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1701831663892/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
jiter==0.5.0
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.5
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langcodes==3.4.0
langsmith==0.1.116
language_data==1.2.0
lazy_loader==0.4
linkify-it-py==2.0.3
lit==18.1.7
llava==0.0.1.dev0
marisa-trie==1.2.0
markdown-it-py==2.2.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mmengine==0.10.4
mmgpt==0.0.1
modelscope==1.16.1
monotonic==1.6
more-itertools==10.4.0
moviepy==2.1.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.5.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
open_clip_torch==2.26.1
openai==1.44.0
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.64
optimum==1.21.2
orjson==3.10.6
oss2==2.19.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.0.1
paramiko==3.4.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
peft==0.12.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.10.0
preshed==3.0.9
proglog==0.1.10
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
proto-plus==1.24.0
protobuf==5.29.2
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1728965148014/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyNaCl==1.5.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1728642231864/work
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rouge==1.0.1
rpds-py==0.20.0
rsa==4.9
ruff==0.6.1
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.13.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.2
spacy-curated-transformers==0.2.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy_pkuseg==1.0.0
SQLAlchemy==2.0.34
srsly==2.4.8
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.37.2
sympy==1.12.1
tenacity==8.5.0
termcolor==2.4.0
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.6.18
tiktoken==0.7.0
timm==0.4.12
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.0
torch==2.0.0
torch-struct==0.5
torchaudio==2.0.1
torchvision==0.15.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1724955920300/work
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
-e git+ssh://git@github.com/RachelWolowitz/MLLM.git@325814761316f39e5632e8b93796b1c4f0357844#egg=transformers&subdirectory=transformers-4.29.2
transformers-stream-generator==0.0.5
triton==2.0.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.17.7
wasabi==1.1.3
watchfiles==0.22.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
weasel==0.4.1
webdataset==0.2.48
websockets==11.0.3
wrapt==1.16.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zh_core_web_sm @ file:///home/mlsnrs/data/wyn/CLIP/MMVP/zh_core_web_sm-3.8.0-py3-none-any.whl#sha256=7de3bd267176b9b2a8defb6997c1cd296da16c57b5e712f72ea44a51755421c8
zipp==3.19.2
